telegram_bot/requirements.txt
requests>=2.28.0
flask>=3.0.0
schedule>=1.2.0
tzdata>=2024.1
python-dotenv>=1.0.0
